\name{lazybear-package}
\alias{lazybear-package}
\alias{lazybear}
\docType{package}
\title{\packageTitle{lazybear}}
\description{\packageDescription{lazybear}}
\section{Package Content}{\packageIndices{lazybear}}
\author{\packageAuthor{lazybear}}
\section{Maintainer}{\packageMaintainer{lazybear}}
\keyword{package}
